+++ books.json
@@ -1,1 +1,1 @@
-[{"publish_date":null,"archived":null,"publisher":null,"cover_state":null,"index":0}]
+[{"publish_date":null,"archived":null,"publisher":null,"cover_state":null,"index":0},{"publish_date":"9999","archived":false,"publisher":"aziz","cover_state":null,"index":1},{"publish_date":"2023","archived":true,"publisher":"abdo","cover_state":null,"index":2}]
