+++ label.json
@@ -1,1 +1,1 @@
-[]
+[{"title":"Graet","color":"Blue","index":0},{"title":"Beayty","color":"Blue","index":1}]
